% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_classFunctions.R
\name{print.openEBGM}
\alias{print.openEBGM}
\title{Print an openEBGM object}
\usage{
\method{print}{openEBGM}(x, threshold = 2, ...)
}
\arguments{
\item{x}{An openEBGM object constructed by \code{ebScores()}}

\item{threshold}{A numeric value indicating the minimum threshold
for QUANT or EBGM values to display.}

\item{...}{Arguments to be passed to other methods}
}
\description{
Print an openEBGM object
}
